Type: Feature
Add a retrofit interface responsible for our retrofit object that gets the network data.

Create an interface that has a method that return a List of our NetworkEntities which is the data model that we just built, inside the GET() annotation you need to write the path where we can find the data we are modelling like a section on the url where we can find the data.